+++ R/packages.R
@@ -4,18 +4,36 @@
 
 .datatable.aware <- TRUE
 
-
-
-# Prevent R CMD check from complaining about the use of pipe expressions
-# standard data.table variables
+# Prevent R CMD check from complaining the standard data.table variables
 if (getRversion() >= "2.15.1") {
   utils::globalVariables(
     c(
+      ".",
+      "GR50",
+      "IC50",
+      "GR Inf",
+      "GR 0",
+      "GEC50",
+      "h GR",
+      "E Inf",
+      "E0",
+      "EC50",
+      "h RV",
+      "GR Max",
+      "E Max",
+      "metric",
+      "Response",
+      ".N",
+      "maxlog10Concentration",
+      "MaxEffectiveness",
+      "RV AOC within set range",
+      "GR AOC within set range",
+      "project_id",
       "rId",
       "cId",
       "concs",
       "type",
       "name"
-    ), 
+    ),
     utils::packageName())
 }